refactor db_util
