+++ .pre-commit-config.yaml
@@ -8,17 +8,3 @@ repos:
     -   id: end-of-file-fixer
     -   id: check-yaml
     -   id: check-added-large-files
--   repo: https://github.com/psf/black
-    rev: 22.3.0
-    hooks:
-    -   id: black
--   repo: local
-    hooks:
-      - id: pylint
-        name: pylint
-        entry: pylint
-        language: system
-        types: [python]
-        exclude: ^experiment_scheduler/master/grpc_master/
-        args:
-          - --max-line-length=120
