+++ experiment_scheduler/db_util/main.py
@@ -1,5 +1,2 @@
-from experiment_scheduler.db_util.experiment import Experiment
-from experiment_scheduler.db_util.task import Task
-
 if __name__ == '__main__':
     pass
